yoloe-service/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0

# HTTP requests
requests==2.31.0
aiofiles==23.2.1

# Computer vision and AI
ultralytics==8.0.206
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3

# Image/video processing
imageio==2.31.6
imageio-ffmpeg==0.4.9

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Optional GPU support
# nvidia-ml-py3==7.352.0  # Uncomment for GPU monitoring

